✨ Refactor repository structure: move database-related logic to dedicated repository modules
