+++ services/catalog/src/simcore_service_catalog/core/events.py
@@ -13,7 +13,7 @@ from servicelib.fastapi.prometheus_instrumentation import (
 
 from .._meta import APP_FINISHED_BANNER_MSG, APP_STARTED_BANNER_MSG
 from ..api.rpc.routes import rpc_api_lifespan
-from ..db.events import database_lifespan
+from ..repository.events import database_lifespan
 from ..services.director import director_lifespan
 from ..services.function_services import function_services_lifespan
 from ..services.rabbitmq import rabbitmq_lifespan
